requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
Django==4.2.5
django-dump-load-utf8==0.0.4
django-filter==23.3
django-jazzmin==2.6.0
djangorestframework==3.14.0
docx2txt==0.8
exceptiongroup==1.1.3
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
Pillow==10.0.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wsproto==1.2.0
